src/main/java/org/gd/leetcode/p0669/Solution.java
package org.gd.leetcode.p0669;

import org.gd.leetcode.common.LeetCode;
import org.gd.leetcode.common.TreeNode;

/**
 * https://leetcode.com/problems/trim-a-binary-search-tree/
 *
 * Given a binary search tree and the lowest and highest boundaries as L and R, trim the tree so that all its elements
 * lies in [L, R] (R >= L). You might need to change the root of the tree, so the result should return the new root of
 * the trimmed binary search tree.
 */
@LeetCode(
        difficulty = LeetCode.Level.EASY,
        tags = LeetCode.Tags.TREE)
class Solution {

    public TreeNode trimBST(TreeNode root, int L, int R) {
        if (root == null)
            return null;

        if (root.val > R)
            return trimBST(root.left, L, R);

        if (root.val < L)
            return trimBST(root.right, L, R);

        root.left = trimBST(root.left, L, R);
        root.right = trimBST(root.right, L, R);
        return root;
    }
}
